Use int instead of long to keep CSV row number (#1287)
* Use int instead of long to keep CSV row number

Generate TooManyEntriesNotice (error) for files that have more than
MAX_INT rows. Such large files would require too much memory and cause
OOM.

* Reduce max ros number to 1 billion and rename TooManyEntries to TooManyRows

We are counting the amount of rows, not entities. A CSV file may have
empty rows that have no GTFS entities.

+++ core/src/main/java/org/mobilitydata/gtfsvalidator/notice/InvalidDateNotice.java
@@ -27,12 +27,11 @@ package org.mobilitydata.gtfsvalidator.notice;
  */
 public class InvalidDateNotice extends ValidationNotice {
   private final String filename;
-  private final long csvRowNumber;
+  private final int csvRowNumber;
   private final String fieldName;
   private final String fieldValue;
 
-  public InvalidDateNotice(
-      String filename, long csvRowNumber, String fieldName, String fieldValue) {
+  public InvalidDateNotice(String filename, int csvRowNumber, String fieldName, String fieldValue) {
     super(SeverityLevel.ERROR);
     this.filename = filename;
     this.csvRowNumber = csvRowNumber;
